requirements.txt
pyscreenshot==3.1
python-dotenv==1.0.0
replicate==0.22.0
Requests==2.31.0
